Formatting, root friendly and environment resilient

+++ scripts/clean-disk.sh
@@ -1,33 +1,53 @@
-echo DISK CLEANUP ============================================
+#!/usr/bin/env bash
 
-echo Before -----------------------
-sudo df -H
+[[ $EUID != 0 ]] && runElevated=sudo
+
+divider() {
+        dividerCharacter=$1
+        printf "%0.s$dividerCharacter" {1..70}
+        printf "\n"
+}
+
+echo
+divider '='
+echo Disk Cleanup
 echo
 
-echo -- Cleaning up unused packages...
-sudo apt autoremove
+divider '-'
+echo Disk Usage: BEFORE
+$runElevated df -H | grep -v '^none'
+echo
+
+echo Cleaning up unused packages...
+$runElevated apt autoremove
 echo
 
 echo Removing unused cached packages...
-sudo apt autoclean
-sudo apt clean
+$runElevated apt autoclean
+$runElevated apt clean
 echo
 
 echo Removing all but the last 3 days of system logs...
-sudo journalctl --vacuum-time=3d
+$runElevated journalctl --vacuum-time=3d
 echo
 
-echo Removing old Snap versions...
-sudo snap set system refresh.retain=2
-set -eu LANG=en_US.UTF-8
-snap list --all | awk '/disabled/{print $1, $3}' | while read snapname revision; do sudo snap remove "$snapname" --revision="$revision"; done
+if command -v snap >/dev/null 2>&1; then
+        echo Removing old Snap versions...
+        $runElevated snap set system refresh.retain=2
+        set -eu LANG=en_US.UTF-8
+        snap list --all | awk '/disabled/{print $1, $3}' | while read snapname revision; do sudo snap remove "$snapname" --revision="$revision"; done
+        echo
+fi
+
+if command -v docker >/dev/null 2>&1; then
+        echo Docker cleanup...
+        docker system prune
+        echo
+fi
+
+divider '-'
+echo Disk Usage: AFTER
+$runElevated df -H | grep -v '^none'
 echo
 
-echo Docker cleanup...
-docker system prune
-echo
-
-echo After ------------------------
-sudo df -H
-echo
-echo DISK CLEANUP \| DONE! ==================================
+echo 'DISK CLEANUP DONE!'
